Hago experimento inicial para sondear la situación #117

+++ Biblioteca-Redes-Neuronales/src/weight-initializer-algorithm/utils.jl
@@ -8,7 +8,7 @@ Comprueba que todos los vectores de `point_dict` no sean ortogonales al vector `
 Esto es que `p.(v - new_point) neq 0` para todo (_,v) en `point_dict`
 `point_dict`` es un diccionario donde los vectores son los valores.
 """
-function notOrtonormal(points::Vector{Vector{Float64}}, p::Vector, new_point::Vector, tam::Int8)::Bool
+function notOrtonormal(points::Vector{Vector{Float64}}, p::Vector, new_point::Vector, tam::Int)::Bool
     for i in 1:tam   
         if sum(p.*(points[i]-new_point)) == 0
             return false
